+++ DependencyInjection/Configuration.php
@@ -22,39 +22,6 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('c975_l_purchase_credits');
-
-        $rootNode
-            ->children()
-                ->arrayNode('creditsNumber')
-                    ->prototype('scalar')->end()
-                ->end()
-                ->arrayNode('creditsPrice')
-                    ->prototype('scalar')->end()
-                ->end()
-                ->booleanNode('gdpr')
-                    ->defaultTrue()
-                ->end()
-                ->scalarNode('currency')
-                    ->defaultValue('EUR')
-                ->end()
-                ->floatNode('vat')
-                    ->defaultNull()
-                ->end()
-                ->scalarNode('live')
-                    ->defaultFalse()
-                ->end()
-                ->scalarNode('userEntity')
-                ->end()
-                ->scalarNode('roleNeeded')
-                    ->defaultValue('ROLE_ADMIN')
-                ->end()
-                ->scalarNode('tosUrl')
-                ->end()
-                ->scalarNode('tosPdf')
-                ->end()
-            ->end()
-        ;
 
         return $treeBuilder;
     }