+++ ChangeLog.md
@@ -1,6 +1,12 @@
 # Changelog
 
+- Created branch 1.x (31/08/2018)
 - Updated composer.json (01/09/2018)
+- Updated `README.md` (01/09/2018)
+
+
+v1.x
+====
 
 v1.6.2
 ------