+++ ChangeLog.md
@@ -3,6 +3,7 @@
 - Created branch 1.x (31/08/2018)
 - Updated composer.json (01/09/2018)
 - Updated `README.md` (01/09/2018)
+- Added `bundle.yaml` (01/09/2018)
 
 
 v1.x