+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Updated `README.md` (01/09/2018)
 - Added `bundle.yaml` (01/09/2018)
 - Added `UPGRADE.md` (01/09/2018)
+- Removed declaration of parameters in Configuration class as they are end-user parameters and defined in c975L/ConfigBundle (01/09/2018)
 
 
 v1.x